docs: fix broken links (#76)
* README.md: fix nits

Fix a couple broken links in the documentation

* README.md: Fix broken links

Fixed many broken links in the doc. Most importantly, the Client
API doc no longer exists, so I pointed that to the Kubebuilder
book, since that provides an interesting insight into Reconcilers,
 Controller, Events and such.

* README.md: Fix Client API doc link

Fix the client API doc link which was previously changed to the
Kubebuilder book, in my previous commit.